refactor(analysis): split 1102-line analysis.ts under 500-LOC limit
The user mandated no file > 500 LOC. analysis.ts had 12 methods
on AnalysisCommands totaling 1102 lines. Split into focused
modules under analysis/:

  - commands/analysis/llm.ts       (164 LOC)  saveLlmAnalysis,
                                              getLlmAnalysis
  - commands/analysis/lifecycle.ts (266 LOC)  seal, rollback, verify,
                                              semanticVerify
  - commands/analysis/stats.ts     (273 LOC)  stats, diff

The AnalysisCommands class stays as the public façade — each
extracted method becomes a 1-line delegator
(`return saveLlmAnalysis(...args)`). External callers
(register.ts, commands.ts) keep working unchanged.

Side effect of the extraction: every method now uses the
`requireProject` guard from #313 directly instead of the inherited
`this.ensureProjectInit + configManager.getProjectId` pair.

Result: analysis.ts shrunk from 1102 → 498 LOC. Under the limit.

Behavior unchanged:
  - typecheck clean
  - biome clean
  - 1029 tests pass (3477 expects)
  - build succeeds
